refactor(engineer_loop): split tests_review_persist.rs (#1266)
tests_review_persist.rs 593 -> 303 LOC; new tests_review_persist_extra.rs.
